Use ngeo-layercatalog directive in layercatalog example
