+++ examples/layercatalog.html
@@ -13,13 +13,20 @@
         width: 600px;
         height: 400px;
       }
+      app-layercatalog button {
+        background: none;
+        border: 0;
+      }
+      app-layercatalog button:hover {
+        opacity: 0.7;
+      }
     </style>
   </head>
-  <body ng-controller="MainController as ctrl">
-    <div id="map" ngeo-map="ctrl.map"></div>
-    <app-catalog app-catalog-map="ctrl.map"></app-catalog>
-    <p>This example shows how to create a Layer catalog with checkboxes to control the addition/removal of layers to/from the map. The catalog is implemented as a recursive directive. The example also shows how to use Angular's one-time binding functionality (with `::` in expressions).</p>
+  <body ng-controller="MainController as mainCtrl">
+    <div id="map" ngeo-map="mainCtrl.map"></div>
+    <app-layercatalog app-layercatalog-map="::mainCtrl.map"></app-layercatalog>
+    <p>This example shows how to use ngeo's layer catalog directive (<code>ngeo-layercatalog</code>). It also shows how to define an application-specific HTML partial for the catalog nodes.</p>
     <script src="../node_modules/angular/angular.js"></script>
-    <script src="/@?main=layer-catalog.js"></script>
+    <script src="/@?main=layercatalog.js"></script>
   </body>
 </html>
