--- examples/partials/catalog.html
@@ -1,4 +1,0 @@
-<ul>
-  <li ng-repeat="node in ::ctrl.tree.children"
-      app-catalog-node="::node" app-catalog-node-map="::ctrl.map"></li>
-</ul>
